requirements.txt
mysql-connector-python==8.0.28
Flask==2.0.2
Werkzeug==2.0.2
python-json-logger==2.0.1
json-log-formatter==0.5.2
